refactor: use state out from server
